🎉 It's ready.... I think
As far as I can detect (its after school hours rn) Timetaber is functioning on a primitive basis featuring the current course in session and the ability to start and stop terms.
Our next steps will be to finalise the next course display, and further forward to create the list view for the day.
All in all, I am extremely satisfied and proud of this creation and will continue to work on Timetaber this year.
Thanks to all my friends and supporters throughout my coding journey, not a bit of which goes to the DoE!
Thank you all! Goodbye!