New ListView API
All ListView values are now stored in the Course structure under `listName` and `listIcon`  instead of redefining a new View for each course (which uses a good 10 lines per course)
To get a ListView, call the `getListView` function with a course and a start time, and it will return a View relative to the Course.

Aditionally, renamed the `globalStorage` class to `storage` and moved it to StorageManager.swift, removed all commented-out older code, cleaned up unused variables and went around trying to improve the efficiency ratio of the code.

Timetaber is currently running error-free.